Rename module `sys-framework` -> `webapp-core`.
