+++ src/webapp-builtins/export/MemoryMonitor.js
@@ -7,7 +7,7 @@ import { Threadlet } from '@this/async';
 import { WallClock } from '@this/clocks';
 import { Duration, Moment } from '@this/data-values';
 import { Host } from '@this/host';
-import { BaseService } from '@this/sys-framework';
+import { BaseService } from '@this/webapp-core';
 import { MustBe } from '@this/typey';
