+++ src/webapp-builtins/export/PathRouter.js
@@ -4,7 +4,7 @@
 import { TreePathKey, TreePathMap } from '@this/collections';
 import { Names } from '@this/compote';
 import { DispatchInfo, IntfRequestHandler, UriUtil } from '@this/net-util';
-import { BaseApplication } from '@this/sys-framework';
+import { BaseApplication } from '@this/webapp-core';
 import { MustBe } from '@this/typey';
