+++ src/webapp-builtins/export/RateLimiter.js
@@ -5,7 +5,7 @@ import { TokenBucket } from '@this/async';
 import { Frequency } from '@this/data-values';
 import { IntfLogger } from '@this/loggy-intf';
 import { IntfRateLimiter } from '@this/net-protocol';
-import { BaseService } from '@this/sys-framework';
+import { BaseService } from '@this/webapp-core';
 import { MustBe } from '@this/typey';
 
 import { RateLimitedStream } from '#p/RateLimitedStream';
