+++ src/webapp-builtins/export/Redirector.js
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { HttpUtil, OutgoingResponse, UriUtil } from '@this/net-util';
-import { BaseApplication } from '@this/sys-framework';
+import { BaseApplication } from '@this/webapp-core';
 import { MustBe } from '@this/typey';
