+++ src/webapp-builtins/export/SerialRouter.js
@@ -3,7 +3,7 @@
 
 import { Names } from '@this/compote';
 import { IntfRequestHandler } from '@this/net-util';
-import { BaseApplication } from '@this/sys-framework';
+import { BaseApplication } from '@this/webapp-core';
 import { MustBe } from '@this/typey';
