+++ src/webapp-builtins/export/SimpleResponse.js
@@ -7,7 +7,7 @@ import { WallClock } from '@this/clocks';
 import { Paths, Statter } from '@this/fs-util';
 import { EtagGenerator, HttpUtil, MimeTypes, OutgoingResponse }
   from '@this/net-util';
-import { BaseApplication } from '@this/sys-framework';
+import { BaseApplication } from '@this/webapp-core';
 import { MustBe } from '@this/typey';
