+++ src/webapp-core/README.md
@@ -1,8 +1,9 @@
-@this/sys-framework
-===================
+@this/webapp-core
+=================
 
-This module contains the Lactoserv framework, including all the base classes
-and plumbing required for the system.
+This module contains the core implementation of the web application server
+framework, including all the base classes and plumbing required to construct
+web applications.
 
 This module notably does _not_ contain any concrete application or service
 implementations. The built-in ones of those live in `@this/webapp-builtins`.
