+++ src/webapp-core/export/BaseApplication.js
