+++ src/webapp-core/export/ComponentManager.js
