+++ src/webapp-core/export/NetworkEndpoint.js
