+++ src/webapp-core/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@this/sys-framework",
+  "name": "@this/webapp-core",
   "version": "0.0.1",
   "type": "module",
   "private": true,
