+++ src/webapp-core/private/ThisModule.js
