+++ src/webapp-util/export/BaseFileService.js
@@ -7,7 +7,7 @@ import { WallClock } from '@this/clocks';
 import { BaseConfig } from '@this/compote';
 import { Duration } from '@this/data-values';
 import { Paths, Statter } from '@this/fs-util';
-import { BaseService } from '@this/sys-framework';
+import { BaseService } from '@this/webapp-core';
 import { MustBe } from '@this/typey';
