+++ src/webapp-util/package.json
@@ -19,7 +19,7 @@
     "@this/fs-util": "*",
     "@this/host": "*",
     "@this/loggy-intf": "*",
-    "@this/sys-framework": "*",
+    "@this/webapp-core": "*",
     "@this/typey": "*"
   }
 }
